+++ test/integration/smtpEmail.test.ts
@@ -1,20 +1,12 @@
 import { expect } from 'chai'
 import express from 'express'
-import { afterEach, beforeEach, describe, it } from 'mocha'
+import { describe, it } from 'mocha'
 
 import http from 'http'
-import { container } from 'tsyringe'
 import { z } from 'zod'
-import { Env } from '../../src/env.js'
-import { resetContainer } from '../../src/ioc.js'
-import createHttpServer from '../../src/server.js'
 import VeritableCloudagentEvents from '../../src/services/veritableCloudagentEvents.js'
-import { cleanupCloudagent } from '../helpers/cloudagent.js'
 import { withCompanyHouseMock } from '../helpers/companyHouse.js'
-import { cleanup } from '../helpers/db.js'
-import { validCompanyNumber } from '../helpers/fixtures.js'
-import { post } from '../helpers/routeHelper.js'
-import { delay } from '../helpers/util.js'
+import { setupSmtpTestEnvironment } from '../helpers/smtp.js'
 
 const ToSchema = z.array(z.string())
 
@@ -39,31 +31,7 @@ describe('SMTP email', () => {
   withCompanyHouseMock()
 
   describe('create invitation and check it has been registered in the email server (happy path)', function () {
-    beforeEach(async () => {
-      process.env.EMAIL_TRANSPORT = 'SMTP_EMAIL'
-      resetContainer()
-      container.registerInstance(Env, new Env())
-      await cleanup()
-      await cleanupCloudagent()
-      server = await createHttpServer()
-      await post(server.app, '/connection/new/create-invitation', {
-        companyNumber: validCompanyNumber,
-        email: 'alice@example.com',
-        action: 'submit',
-      })
-
-      // Allow time for the email to be sent
-      await delay(2000)
-    })
-
-    afterEach(async () => {
-      await cleanup()
-      await clearSmtp4devMessages()
-      await cleanupCloudagent()
-      server.cloudagentEvents.stop()
-      process.env.EMAIL_TRANSPORT = 'STREAM'
-      resetContainer()
-    })
+    server = setupSmtpTestEnvironment(server)
 
     it('should send an email via SMTP', async () => {
       const options = {
@@ -119,28 +87,4 @@ describe('SMTP email', () => {
       await emailCheck
     })
   })
-
-  async function clearSmtp4devMessages() {
-    return new Promise((resolve, reject) => {
-      const options = {
-        hostname: 'localhost',
-        port: 5000,
-        path: '/api/messages/*', // Delete all messages
-        method: 'DELETE',
-        headers: {
-          'Content-Type': 'application/json',
-        },
-      }
-
-      const req = http.request(options, (res) => {
-        if (res.statusCode === 200) {
-          resolve(true)
-        } else {
-          reject(new Error(`Failed to clear messages, status code: ${res.statusCode}`))
-        }
-      })
-      req.on('error', (error) => reject(error))
-      req.end()
-    })
-  }
 })
